+++ ChangeLog.md
@@ -1,9 +1,58 @@
 # ChangeLog
 
+## [1.0.4] - 2024-10-8
+
+### Added
+- self.effort and wm.see_time are added
+
+### Fixed
+-
+
+### Changed
+- 
+
+### Developer 
+- [SoroushMazloum](https://github.com/SoroushMazloum)
+
+
+## [1.0.3] - 2024-10-7
+
+### Added
+- penalty_kick_state has been added to the proxy in the WorldModel message.
+
+### Fixed
+-
+
+### Changed
+- 
+
+### Developer
+- [SadraKhanjari](https://github.com/SK2iP)
+- [SoroushMazloum](https://github.com/SoroushMazloum)
+
+
+## [1.0.2] - 2024-09-15
+
+### Added
+- ignore_doforcekick and ignore_doHeardPassRecieve has been added to the proxy in the actions message.
+- now users can decide to do the doForceKick and doHeardPassRecieve or not
+
+### Fixed
+- now state only generate once per cycle 
+
+### Changed
+- 
+
+### Developer
+- [NaderZare](https://github.com/naderzare)
+- [SadraKhanjari](https://github.com/SK2iP)
+
 ## [1.0.1] - 2024-09-15
 
 ### Added
 - catch_time has been added to the proxy in the self message.
+- kickable_opponent_existance and kickable_teammate_existance has been added to the proxy in the worldmodel message.
+- bhv_doforceKick action has been added as a message and to the actions message.
 
 ### Fixed
 - 
@@ -11,8 +60,9 @@
 ### Changed
 - 
 
-### Engineers
-- [SoroushMazloum](https://github.com/)
+### Developer
+- [SoroushMazloum](https://github.com/SoroushMazloum)
+- [SadraKhanjari](https://github.com/SK2iP)
 
 ## [1.0.0] - 2024-09-15
 
@@ -27,7 +77,7 @@
 ### Changed
 - changed chain_action messages name to planner
 
-### Engineers
+### Developer
 - [NaderZare](https://github.com/naderzare)
 - [SadraKhanjari](https://github.com/SK2iP)